+++ Models/ResponseModels/Transactions/VendorInvoiceTxn/VendorInvoiceTxnReadResponseModel.cs
@@ -27,12 +27,12 @@ namespace Models.ResponseModels.VendorInvoiceTxn.VendorInvoiceTxn
 
         // Tab2
         public List<InvoiceFessDetailList>? invoiceFeeDetails { get; set; }
-        public int? professionalFeeAmt { get; set; }
-        public int? govtOrOfficialFeeAmt { get; set; }
-        public int? otherChargesAmt { get; set; }
-        public int? discountAmt { get; set; }
-        public int? discountCreditNoteAmt { get; set; }
-        public int? totalAmount { get; set; }
+        public decimal? professionalFeeAmt { get; set; }
+        public decimal? govtOrOfficialFeeAmt { get; set; }
+        public decimal? otherChargesAmt { get; set; }
+        public decimal? discountAmt { get; set; }
+        public decimal? discountCreditNoteAmt { get; set; }
+        public decimal? totalAmount { get; set; }
         public string? postedInTally { get; set; }
         public string? patentNo { get; set; }
         public string? creditNoteNo { get; set; }
@@ -50,7 +50,7 @@ namespace Models.ResponseModels.VendorInvoiceTxn.VendorInvoiceTxn
         // Tab 4
         public string? customerPONo { get; set; }
         public DateTime? pODate { get; set; }
-        public int? pOValueInclusiveTaxes { get; set; }
+        public decimal? pOValueInclusiveTaxes { get; set; }
         public string? saleCurrency { get; set; }
         public List<InvoiceFessDetailList>? feeDetails { get; set; }
         public List<SaleInvoiceDetailList>? saleDetails { get; set; }
